Corrects numpy bug for heroku

+++ app.py
@@ -152,7 +152,7 @@ if choose in ["Dettaglio", "Wordcloud", "Recensioni"]:
                     "ANNO",
                     venues_df.Data.dt.year.min(),
                     venues_df.Data.dt.year.max(),
-                    (float(venues_df.Data.dt.year.min()), float(venues_df.Data.dt.year.max()))
+                    (int(venues_df.Data.dt.year.min()), int(venues_df.Data.dt.year.max()))
                 )
             start_date = start_date = pd.to_datetime(str(year[0]), format="%Y")
             end_date = pd.to_datetime(str(year[1]), format="%Y")
